feat(tailwind): add compile-time validation for missing css classes
This introduces early verification of Tailwind classes directly in the  macro during build. It guarantees that any misspelled or undefined class used in  or  throws a compile-time error. Also refactored  into a module for better maintainability.
