+++ crates/vertigo-macro/src/trace_tailwind/mod.rs
@@ -1,0 +1,7 @@
+mod bundle;
+mod macro_impl;
+pub(crate) mod paths;
+pub(crate) mod validate;
+
+pub(crate) use bundle::bundle_tailwind;
+pub(crate) use macro_impl::{add_to_tailwind, trace_tailwind};
